+++ CHANGELOG.md
@@ -125,6 +125,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Support resolving `Order` as an entity in Apollo Federation - #12328 by @binary-koan
 - [Preview] Add `ProductBulkCreate` mutation - #12177 by @SzymJ
 - [Preview] Add `CustomerBulkUpdate` mutation - #12268 by @SzymJ
+- [Preview] Add `OrderBulkCreate` mutation - #12269 by @zedzior
 
 ### Other changes
 - Create order discounts for all voucher types - #12272 by @IKarbowiak